+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/CustomTabBrowserFragment.kt
@@ -26,6 +26,7 @@ class CustomTabBrowserFragment : BaseBrowserFragment(), BackHandler {
                 layout.toolbar,
                 sessionId,
                 components.menuBuilder,
+                window = activity?.window,
                 closeListener = { activity?.finish() }),
             owner = this,
             view = layout)